ui: adopt Form.Viewport in cards and dialogs; unblock react-refresh
Form-in-card workarounds: SegmentCard, CreateObjectPanel and ChannelCreatePanel
each hand-placed a form to dodge the nested-gutter bug that useInColumn now
handles, so all three use Form.Viewport. Verified in storybook: SegmentCard's
title and its form labels share one edge at 41px with zero nested column roots,
and the create dialog's labels, inputs and save button all sit at 33px inside a
384px dialog.

MarkdownProperties: the current-branch row is a Form.Label with the branch and
checkpoint counts as labelEnd, so it takes the 32px field-label geometry instead
of a hand-rolled flex row.

AwaitingObject issued Query.select(Filter.everything()) against client.spaces.
Unlike db.query, the graph-level queryable does not inject a from() clause, so
the planner threw "Query must be scoped with a from() clause" and killed
whatever had rendered the toast (intermittently fatal for the MarkdownArticle
story). It now scopes the query to the space dbs.

Fast refresh: DensityProvider, Listbox, OrderedListRoot, AccordionItem and
Logger each exported a context, hook or helper beside their components, which
disables react-refresh for the module — every edit full-reloaded the app and
cost a 30s Composer restart. Those exports move to sibling modules and are
re-exported from the barrels. Edits to the list components now log an hmr update
with no page reload. Other blockers remain (a react-client circular import, a
plugin-settings default export).
