+++ packages/plugins/plugin-review/src/components/MarkdownProperties/MarkdownProperties.tsx
@@ -41,13 +41,17 @@ export const MarkdownProperties = ({ subject }: MarkdownPropertiesProps) => {
 
   return (
     <Form.Section title={t('versions.title')}>
-      <div className='flex items-center gap-2 pli-1 text-sm'>
-        <span className='truncate'>{currentLabel}</span>
-        <span className='ms-auto shrink-0 text-xs text-description'>
-          {t('branch-count.label', { count: branchCount })} · {t('checkpoint-count.label', { count: versionCount })}
-        </span>
-      </div>
-      <div className='flex gap-1 pli-1'>
+      {/* `standalone` labels nothing focusable, so it renders a span rather than an orphan <label>. */}
+      <Form.Label
+        standalone
+        label={currentLabel}
+        labelEnd={
+          <span className='shrink-0 text-xs text-description'>
+            {t('branch-count.label', { count: branchCount })} · {t('checkpoint-count.label', { count: versionCount })}
+          </span>
+        }
+      />
+      <div className='flex gap-1'>
         <IconButton
           icon='ph--bookmark-simple--regular'
           label={t('create-checkpoint.label')}
