+++ packages/plugins/plugin-space/src/components/CreateObjectPanel/CreateObjectPanel.tsx
@@ -8,7 +8,7 @@ import { type Collection, type Database, Obj } from '@dxos/echo';
 import { type AnyProperties } from '@dxos/echo/internal';
 import { type SpaceId } from '@dxos/keys';
 import { type Space } from '@dxos/react-client/echo';
-import { Column, Icon, toLocalizedString, useDefaultValue, useTranslation } from '@dxos/react-ui';
+import { Icon, toLocalizedString, useDefaultValue, useTranslation } from '@dxos/react-ui';
 import { Form, omitId } from '@dxos/react-ui-form';
 import { Picker } from '@dxos/react-ui-list';
 import { SearchList, useSearchListResults } from '@dxos/react-ui-search';
@@ -115,27 +115,23 @@ export const CreateObjectPanel = ({
   }
 
   if (metadata.inputSchema) {
-    // The host (Dialog.Body) already owns the gutter Column. Use Column.Center to place the form in
-    // the same center column as the dialog title — NOT a subgrid wrapper or Form.Viewport's own
-    // Column.Root. Subgrid can't propagate through Form.Root's `display: contents` wrapper (the form
-    // would fall back to a stray track and inset), and a nested Column.Root would double-gutter it.
     return (
-      <Column.Center>
-        <Form.Root
-          autoFocus
-          schema={inputSchema}
-          defaultValues={initialFormValues}
-          fieldProvider={inputSurfaceLookup}
-          db={Obj.isObject(target) ? Obj.getDatabase(target) : target}
-          onSave={handleCreateObject}
-          testId='create-object-form'
-        >
+      <Form.Root
+        autoFocus
+        schema={inputSchema}
+        defaultValues={initialFormValues}
+        fieldProvider={inputSurfaceLookup}
+        db={Obj.isObject(target) ? Obj.getDatabase(target) : target}
+        onSave={handleCreateObject}
+        testId='create-object-form'
+      >
+        <Form.Viewport>
           <Form.Content>
             <Form.FieldSet />
             <Form.Submit />
           </Form.Content>
-        </Form.Root>
-      </Column.Center>
+        </Form.Viewport>
+      </Form.Root>
     );
   }
