+++ packages/ui/react-ui-debug/src/components/Logger/index.ts
@@ -3,4 +3,9 @@
 //
 
 export * from './Logger';
+// Re-exported here rather than from `Logger.tsx`: a non-component export in that module disables
+// react-refresh for the whole logger, so every edit full-reloads the app.
+export { copyToClipboard, levelColor, logLevelsAspect, useLoggerContext } from './LoggerContext';
+export { type LogRow, logBuffer } from './log-buffer';
+export { type LevelName, LEVELS, type LogRecorder, composeFilter, startLogRecording } from './recorder';
 export * from './format';
