+++ packages/ui/react-ui-list/src/components/Listbox/Listbox.tsx
@@ -46,7 +46,6 @@
 //   - Virtualization or drag-and-drop. Reach for `@dxos/react-ui-mosaic`.
 //   - Multi-select. Future expansion — the aspect (`useListSelection`) already supports it.
 
-import { createContext } from '@radix-ui/react-context';
 import React, {
   type ComponentPropsWithRef,
   type FocusEvent,
@@ -71,49 +70,25 @@ import {
 } from '@dxos/react-ui';
 import { mx } from '@dxos/ui-theme';
 
-import {
-  type SelectionItemBinding,
-  type UseListSelectionReturn,
-  useListNavigation,
-  useListSelection,
-} from '../../aspects';
+import { type SelectionItemBinding, useListNavigation, useListSelection } from '../../aspects';
 import { listTheme } from '../List.theme';
+import {
+  LISTBOX_ITEM_NAME,
+  type ListboxItemContextValue,
+  ListboxItemProvider,
+  ListboxProvider,
+  useListboxContext,
+  useListboxItemContext,
+} from './ListboxContext';
 import { ListItemContent, type ListItemContentProps } from './ListItemContent';
 
 const styles = listTheme.styles();
 
-const LISTBOX_NAME = 'Listbox';
 const LISTBOX_ROOT_NAME = 'Listbox.Root';
 const LISTBOX_VIEWPORT_NAME = 'Listbox.Viewport';
 const LISTBOX_CONTENT_NAME = 'Listbox.Content';
-const LISTBOX_ITEM_NAME = 'Listbox.Item';
 const LISTBOX_ITEM_LABEL_NAME = 'Listbox.ItemLabel';
 const LISTBOX_INDICATOR_NAME = 'Listbox.Indicator';
-
-//
-// Contexts — plain Radix contexts (un-scoped). Scoped composition (nested Listboxes,
-// Combobox embeddings) is a future expansion; when needed, switch to `createContextScope`
-// and thread `__listboxScope` through every subcomponent's props in one focused PR.
-//
-
-type ListboxContextValue = {
-  /**
-   * Whether the list participates in selection. Inferred on `Root` from the presence of
-   * `value`/`defaultValue`/`onValueChange`. Drives `role` (listbox/option vs list/listitem),
-   * `aria-selected`, and whether row clicks update the selection model.
-   */
-  selectable: boolean;
-  /** Selection aspect binding factory; items consume their own bindings from this. */
-  selection: UseListSelectionReturn;
-};
-
-type ListboxItemContextValue = {
-  id: string;
-  selected: boolean;
-};
-
-const [ListboxProvider, useListboxContext] = createContext<ListboxContextValue>(LISTBOX_NAME);
-const [ListboxItemProvider, useListboxItemContext] = createContext<ListboxItemContextValue>(LISTBOX_ITEM_NAME);
 
 //
 // Root — headless context provider. Renders no DOM.
@@ -384,16 +359,6 @@ const Indicator = forwardRef<SVGSVGElement, IndicatorProps>(({ classNames, ...ro
 
 Indicator.displayName = LISTBOX_INDICATOR_NAME;
 
-/**
- * Read selection state for a single id from inside any descendant of `<Listbox.Root>`.
- * Returns `true` when the row is currently selected. Lets composing components react to
- * selection without re-rendering on unrelated changes.
- */
-const useListboxSelection = (id: string): boolean => {
-  const { selection } = useListboxContext('useListboxSelection');
-  return selection.bind(id).selected;
-};
-
 //
 // Public namespace.
 //
@@ -408,7 +373,7 @@ const Listbox = {
   Indicator,
 };
 
-export { Listbox, useListboxSelection };
+export { Listbox };
 
 export type {
   ListItemContentProps as ItemContentProps,
