+++ packages/ui/react-ui-list/src/components/Listbox/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './Listbox';
+export { useListboxSelection } from './ListboxContext';
