+++ packages/ui/react-ui/src/primitives/DensityProvider/index.ts
@@ -2,4 +2,5 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './DensityContext';
 export * from './DensityProvider';
